+++ test-output/Default suite/Default test.html
@@ -57,9 +57,9 @@ function toggleAllBoxes() {
 <tr>
 <td>Tests passed/Failed/Skipped:</td><td>1/0/0</td>
 </tr><tr>
-<td>Started on:</td><td>Sun Mar 02 17:28:40 IST 2025</td>
+<td>Started on:</td><td>Sun Mar 02 20:17:44 IST 2025</td>
 </tr>
-<tr><td>Total time:</td><td>36 seconds (36145 ms)</td>
+<tr><td>Total time:</td><td>0 seconds (109 ms)</td>
 </tr><tr>
 <td>Included groups:</td><td></td>
 </tr><tr>
@@ -76,10 +76,10 @@ function toggleAllBoxes() {
 <td><b>Instance</b></td>
 </tr>
 <tr>
-<td title='WebDriverActions.WindowHandlingTest.windowHandlingTest()'><b>windowHandlingTest</b><br>Test class: WebDriverActions.WindowHandlingTest</td>
+<td title='WebDriverActions.FileHandlingTest.fileHandlingTest()'><b>fileHandlingTest</b><br>Test class: WebDriverActions.FileHandlingTest</td>
 <td></td>
-<td>35</td>
-<td>WebDriverActions.WindowHandlingTest@402c4085</td></tr>
+<td>0</td>
+<td>WebDriverActions.FileHandlingTest@1a75e76a</td></tr>
 </table><p>
 </body>
 </html>
