+++ test-output/Default suite/Default test.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitXMLReporter -->
-<testsuite ignored="0" hostname="DESKTOP-QR0VDMM" failures="0" tests="1" name="Default test" time="36.145" errors="0" timestamp="2025-03-02T17:29:18 IST">
-  <testcase classname="WebDriverActions.WindowHandlingTest" name="windowHandlingTest" time="35.997"/>
+<testsuite ignored="0" hostname="DESKTOP-QR0VDMM" failures="0" tests="1" name="Default test" time="0.109" errors="0" timestamp="2025-03-02T20:17:44 IST">
+  <testcase classname="WebDriverActions.FileHandlingTest" name="fileHandlingTest" time="0.015"/>
 </testsuite> <!-- Default test -->
