+++ test-output/junitreports/TEST-WebDriverActions.FileHandlingTest.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated by org.testng.reporters.JUnitReportReporter -->
+<testsuite hostname="DESKTOP-QR0VDMM" failures="0" tests="1" name="WebDriverActions.FileHandlingTest" time="0.015" errors="0" timestamp="2025-03-02T20:17:45 IST" skipped="0">
+  <testcase classname="WebDriverActions.FileHandlingTest" name="fileHandlingTest" time="0.015"/>
+</testsuite> <!-- WebDriverActions.FileHandlingTest -->
