+++ test-output/testng-results.xml
@@ -2,16 +2,16 @@
 <testng-results ignored="0" total="1" passed="1" failed="0" skipped="0">
   <reporter-output>
   </reporter-output>
-  <suite started-at="2025-03-02T17:28:40 IST" name="Default suite" finished-at="2025-03-02T17:29:16 IST" duration-ms="36145">
+  <suite started-at="2025-03-02T20:17:44 IST" name="Default suite" finished-at="2025-03-02T20:17:44 IST" duration-ms="109">
     <groups>
     </groups>
-    <test started-at="2025-03-02T17:28:40 IST" name="Default test" finished-at="2025-03-02T17:29:16 IST" duration-ms="36145">
-      <class name="WebDriverActions.WindowHandlingTest">
-        <test-method signature="windowHandlingTest()[pri:0, instance:WebDriverActions.WindowHandlingTest@402c4085]" started-at="2025-03-02T17:28:40 IST" name="windowHandlingTest" finished-at="2025-03-02T17:29:16 IST" duration-ms="35997" status="PASS">
+    <test started-at="2025-03-02T20:17:44 IST" name="Default test" finished-at="2025-03-02T20:17:44 IST" duration-ms="109">
+      <class name="WebDriverActions.FileHandlingTest">
+        <test-method signature="fileHandlingTest()[pri:0, instance:WebDriverActions.FileHandlingTest@1a75e76a]" started-at="2025-03-02T20:17:44 IST" name="fileHandlingTest" finished-at="2025-03-02T20:17:44 IST" duration-ms="15" status="PASS">
           <reporter-output>
           </reporter-output>
-        </test-method> <!-- windowHandlingTest -->
-      </class> <!-- WebDriverActions.WindowHandlingTest -->
+        </test-method> <!-- fileHandlingTest -->
+      </class> <!-- WebDriverActions.FileHandlingTest -->
     </test> <!-- Default test -->
   </suite> <!-- Default suite -->
 </testng-results>
